Add script to convert x86 RPM dependencies to PPC
VSCode specifies dependencies for its RPM package installation in the
`resources/linux/rpm/dependencies.json` file, which currently lists
dependencies for x86_64. The dependencies are listed as versioned
sonames, some of which are not available on PPC64. This script checks
those dependencies and adds them to the `dependencies.json` file.

Signed-off-by: Colin Samples <colin.samples+git@gmail.com>